+++ src/Redis/Command/Explain.php
@@ -21,6 +21,8 @@ declare(strict_types=1);
 
 namespace MacFJA\RediSearch\Redis\Command;
 
+use MacFJA\RediSearch\Redis\Command\Option\CustomValidatorOption as CV;
+use MacFJA\RediSearch\Redis\Command\Option\NamedOption;
 use MacFJA\RediSearch\Redis\Command\Option\NamelessOption;
 
 /**
@@ -33,6 +35,7 @@ class Explain extends AbstractCommand
         parent::__construct([
             'index' => new NamelessOption(null, '>=2.0.0'),
             'query' => new NamelessOption(null, '>=2.0.0'),
+            'dialect' => CV::isNumeric(new NamedOption('DIALECT', null, '>=2.4.3')),
         ], $rediSearchVersion);
     }
 
@@ -46,6 +49,13 @@ class Explain extends AbstractCommand
     public function setQuery(string $query): self
     {
         $this->options['query']->setValue($query);
+
+        return $this;
+    }
+
+    public function setDialect(int $version): self
+    {
+        $this->options['dialect']->setValue($version);
 
         return $this;
     }