+++ src/Redis/Command/Search.php
@@ -77,6 +77,7 @@ class Search extends AbstractCommand implements PaginatedCommand
                 'payload' => new NamedOption('PAYLOAD', null, '>=2.0.0'),
                 'sortby' => new SortByOption(),
                 'limit' => new LimitOption(),
+                'dialect' => CV::isNumeric(new NamedOption('DIALECT', null, '>=2.4.3')),
             ],
             $rediSearchVersion
         );
@@ -299,6 +300,13 @@ class Search extends AbstractCommand implements PaginatedCommand
         $limit = $this->options['limit'];
 
         return $limit->getSize();
+    }
+
+    public function setDialect(int $version): self
+    {
+        $this->options['dialect']->setValue($version);
+
+        return $this;
     }
 
     /**